feat(light): add light

+++ engine/src/classes/modelConfig/modelConfig.cpp
@@ -3,10 +3,10 @@
 ModelConfig::ModelConfig() {}
 
 ModelConfig::ModelConfig(string texture_file_path, string color,
-                         tuple<int, int, int> diffuse,
-                         tuple<int, int, int> ambient,
-                         tuple<int, int, int> specular,
-                         tuple<int, int, int> emissive, int shininess) {
+                         tuple<float, float, float> diffuse,
+                         tuple<float, float, float> ambient,
+                         tuple<float, float, float> specular,
+                         tuple<float, float, float> emissive, float shininess) {
   this->texture_file_path = texture_file_path;
   this->color = color;
   this->diffuse = diffuse;
@@ -20,15 +20,19 @@ string ModelConfig::get_texture_file_path() { return this->texture_file_path; }
 
 string ModelConfig::get_color() { return this->color; }
 
-tuple<int, int, int> ModelConfig::get_diffuse() { return this->diffuse; }
+tuple<float, float, float> ModelConfig::get_diffuse() { return this->diffuse; }
 
-tuple<int, int, int> ModelConfig::get_ambient() { return this->ambient; }
+tuple<float, float, float> ModelConfig::get_ambient() { return this->ambient; }
 
-tuple<int, int, int> ModelConfig::get_specular() { return this->specular; }
+tuple<float, float, float> ModelConfig::get_specular() {
+  return this->specular;
+}
 
-tuple<int, int, int> ModelConfig::get_emissive() { return this->emissive; }
+tuple<float, float, float> ModelConfig::get_emissive() {
+  return this->emissive;
+}
 
-int ModelConfig::get_shininess() { return this->shininess; }
+float ModelConfig::get_shininess() { return this->shininess; }
 
 void ModelConfig::set_texture_file_path(string path) {
   this->texture_file_path = path;
@@ -36,36 +40,38 @@ void ModelConfig::set_texture_file_path(string path) {
 
 void ModelConfig::set_color(string color) { this->color = color; }
 
-void ModelConfig::set_diffuse(tuple<int, int, int> diffuse) {
+void ModelConfig::set_diffuse(tuple<float, float, float> diffuse) {
   this->diffuse = diffuse;
 }
 
-void ModelConfig::set_diffuse(int r, int g, int b) {
+void ModelConfig::set_diffuse(float r, float g, float b) {
   this->diffuse = make_tuple(r, g, b);
 }
 
-void ModelConfig::set_ambient(tuple<int, int, int> ambient) {
+void ModelConfig::set_ambient(tuple<float, float, float> ambient) {
   this->ambient = ambient;
 }
 
-void ModelConfig::set_ambient(int r, int g, int b) {
+void ModelConfig::set_ambient(float r, float g, float b) {
   this->ambient = make_tuple(r, g, b);
 }
 
-void ModelConfig::set_specular(tuple<int, int, int> specular) {
+void ModelConfig::set_specular(tuple<float, float, float> specular) {
   this->specular = specular;
 }
 
-void ModelConfig::set_specular(int r, int g, int b) {
+void ModelConfig::set_specular(float r, float g, float b) {
   this->specular = make_tuple(r, g, b);
 }
 
-void ModelConfig::set_emissive(tuple<int, int, int> emissive) {
+void ModelConfig::set_emissive(tuple<float, float, float> emissive) {
   this->emissive = emissive;
 }
 
-void ModelConfig::set_emissive(int r, int g, int b) {
+void ModelConfig::set_emissive(float r, float g, float b) {
   this->emissive = make_tuple(r, g, b);
 }
 
-void ModelConfig::set_shininess(int shininess) { this->shininess = shininess; }
+void ModelConfig::set_shininess(float shininess) {
+  this->shininess = shininess;
+}
